refactor(ecs) - convert into modular header files
