+++ include/system/ecs.h
@@ -1,0 +1,6 @@
+#pragma once
+#include "ecs/ecs_types.h"
+#include "ecs/resource_manager.h"
+#include "ecs/resource_pool.h"
+#include "ecs/task_manager.h"
+#include "ecs/syscall.h"
