--- include/system/ecs_functions.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "./ecs_types.h"
-
-namespace ECS
-{
-}
