+++ test/ecs_test.cpp
@@ -1,5 +1,12 @@
+#include "system/ecs.h"
 #include "pch.h"
-#include "system/ecs_types.h"
+
+using System::ECS::pid;
+using System::ECS::Syscall;
+using System::ECS::TaskManager;
+using System::ECS::ResourceManager;
+using System::ECS::ResourcePool;
+using System::ECS::SyscallType;
 
 struct test_component
 {
@@ -17,12 +24,12 @@ struct test_component_2
     int value;
 };
 
-void test_system([[maybe_unused]] System::pid id, test_component& comp)
+void test_system([[maybe_unused]] pid id, test_component& comp)
 {
     comp.value += 1;
 }
 
-void test_system_2([[maybe_unused]] System::pid id, test_component& comp)
+void test_system_2([[maybe_unused]] pid id, test_component& comp)
 {
     comp.value *= 2;
 }
@@ -31,9 +38,9 @@ void test_system_2([[maybe_unused]] System::pid id, test_component& comp)
 
 TEST(ECS, add_component_test)
 {
-    using TestResource = System::ResourceManager<1000, test_component>;
+    using TestResource = ResourceManager<1000, test_component>;
     TestResource resource;
-    const System::pid id = resource.add_process();
+    const pid id = resource.add_process();
     resource.add_resource<test_component>(id, test_component{1});
     EXPECT_EQ(resource.query<test_component>().get(id).value, 1);
     EXPECT_TRUE(resource.query<test_component>().has(id));
@@ -41,11 +48,11 @@ TEST(ECS, add_component_test)
 
 TEST(ECS, add_multiple_components_test)
 {
-    using TestResource = System::ResourceManager<1000, test_component>;
+    using TestResource = ResourceManager<1000, test_component>;
     TestResource resource;
-    const System::pid id_1 = resource.add_process();
+    const pid id_1 = resource.add_process();
     resource.add_resource<test_component>(id_1, test_component{1});
-    const System::pid id_2 = resource.add_process();
+    const pid id_2 = resource.add_process();
     resource.add_resource<test_component>(id_2, test_component{2});
     EXPECT_EQ(resource.query<test_component>().get(id_1).value, 1);
     EXPECT_EQ(resource.query<test_component>().get(id_2).value, 2);
@@ -53,9 +60,9 @@ TEST(ECS, add_multiple_components_test)
 
 TEST(ECS, remove_component_test)
 {
-    using TestResource = System::ResourceManager<1000, test_component>;
+    using TestResource = ResourceManager<1000, test_component>;
     TestResource resource;
-    const System::pid id = resource.add_process();
+    const pid id = resource.add_process();
     resource.add_resource<test_component>(id, test_component{1});
     EXPECT_TRUE(resource.query<test_component>().has(id));
     resource.remove_resource<test_component>(id);
@@ -64,9 +71,9 @@ TEST(ECS, remove_component_test)
 
 TEST(ECS, layer_components_test)
 {
-    using TestResource = System::ResourceManager<1000, test_component, test_component_2>;
+    using TestResource = ResourceManager<1000, test_component, test_component_2>;
     TestResource resource;
-    const System::pid id_1 = resource.add_process();
+    const pid id_1 = resource.add_process();
     resource.add_resource<test_component>(id_1, test_component{1});
     resource.add_resource<test_component_2>(id_1, test_component_2{2});
     EXPECT_EQ(resource.query<test_component>().get(id_1).value, 1);
@@ -75,9 +82,9 @@ TEST(ECS, layer_components_test)
 
 TEST(ECS, delete_resource_test)
 {
-    using TestResource = System::ResourceManager<1000, test_component, test_component_2>;
+    using TestResource = ResourceManager<1000, test_component, test_component_2>;
     TestResource resource;
-    const System::pid id_1 = resource.add_process();
+    const pid id_1 = resource.add_process();
     resource.add_resource<test_component>(id_1, test_component{1});
     resource.add_resource<test_component_2>(id_1, test_component_2{2});
     EXPECT_EQ(resource.query<test_component>().get(id_1).value, 1);
@@ -91,10 +98,10 @@ TEST(ECS, delete_resource_test)
 
 TEST(ECS, exec_add_component)
 {
-    using TestResource = System::ResourceManager<1000, test_component>;
+    using TestResource = ResourceManager<1000, test_component>;
     TestResource resource;
-    System::Syscall syscall{resource};
-    const System::pid id_1 = resource.add_process();
+    Syscall syscall{resource};
+    const pid id_1 = resource.add_process();
     syscall.add_component(id_1, test_component{1});
     EXPECT_FALSE(resource.query<test_component>().has(id_1));
     syscall.exec();
@@ -104,10 +111,10 @@ TEST(ECS, exec_add_component)
 
 TEST(ECS, exec_create_entity)
 {
-    using TestResource = System::ResourceManager<1000, test_component>;
+    using TestResource = ResourceManager<1000, test_component>;
     TestResource resource;
-    System::Syscall syscall{resource};
-    const System::pid id_1 = syscall.create_entity(test_component{1});
+    Syscall syscall{resource};
+    const pid id_1 = syscall.create_entity(test_component{1});
     EXPECT_FALSE(resource.query<test_component>().has(id_1));
     syscall.exec();
     EXPECT_TRUE(resource.query<test_component>().has(id_1));
@@ -115,10 +122,10 @@ TEST(ECS, exec_create_entity)
 
 TEST(ECS, exec_remove_component)
 {
-    using TestResource = System::ResourceManager<1000, test_component>;
+    using TestResource = ResourceManager<1000, test_component>;
     TestResource resource;
-    System::Syscall syscall{resource};
-    const System::pid id_1 = resource.add_process();
+    Syscall syscall{resource};
+    const pid id_1 = resource.add_process();
     resource.add_resource<test_component>(id_1, test_component{1});
     syscall.remove_component<test_component>(id_1);
     EXPECT_TRUE(resource.query<test_component>().has(id_1));
@@ -128,10 +135,10 @@ TEST(ECS, exec_remove_component)
 
 TEST(ECS, exec_remove_entity)
 {
-    using TestResource = System::ResourceManager<1000, test_component, test_component_2>;
+    using TestResource = ResourceManager<1000, test_component, test_component_2>;
     TestResource resource;
-    System::Syscall syscall{resource};
-    const System::pid id_1 = resource.add_process();
+    Syscall syscall{resource};
+    const pid id_1 = resource.add_process();
     resource.add_resource<test_component>(id_1, test_component{1});
     resource.add_resource<test_component_2>(id_1, test_component_2{2});
     syscall.remove_entity(id_1);
@@ -146,13 +153,13 @@ TEST(ECS, exec_remove_entity)
 
 TEST(ECS, system_run_test)
 {
-    using TestResource = System::ResourceManager<1000, test_component>;
-    using TestSyscall = System::Syscall<1000, test_component>;
+    using TestResource = ResourceManager<1000, test_component>;
+    using TestSyscall = Syscall<1000, test_component>;
     TestResource resource;
     TestSyscall syscall{resource};
-    const System::pid id_1 = resource.add_process();
+    const pid id_1 = resource.add_process();
     resource.add_resource<test_component>(id_1, test_component{1});
-    System::TaskManager<TestResource, TestSyscall, test_system> task_manager(resource, syscall);
+    TaskManager<TestResource, TestSyscall, test_system> task_manager(resource, syscall);
     EXPECT_EQ(resource.query<test_component>().get(id_1).value, 1);
     task_manager.run_all();
     EXPECT_EQ(resource.query<test_component>().get(id_1).value, 2);
@@ -160,32 +167,32 @@ TEST(ECS, system_run_test)
 
 TEST(ECS, system_sequence_test)
 {
-    using TestResource = System::ResourceManager<1000, test_component>;
-    using TestSyscall = System::Syscall<1000, test_component>;
+    using TestResource = ResourceManager<1000, test_component>;
+    using TestSyscall = Syscall<1000, test_component>;
     TestResource resource;
     TestSyscall syscall{resource};
-    const System::pid id_1 = resource.add_process();
+    const pid id_1 = resource.add_process();
     resource.add_resource<test_component>(id_1, test_component{1});
-    System::TaskManager<TestResource, TestSyscall, test_system, test_system_2> task_manager(resource, syscall);
+    TaskManager<TestResource, TestSyscall, test_system, test_system_2> task_manager(resource, syscall);
     EXPECT_EQ(resource.query<test_component>().get(id_1).value, 1);
     task_manager.run_all();
     EXPECT_EQ(resource.query<test_component>().get(id_1).value, 4); // (1+1)*2
 }
 
-void test_system_3(System::pid id, System::Syscall<1000, test_component> &syscall, [[maybe_unused]] test_component& comp)
+void test_system_3(pid id, Syscall<1000, test_component> &syscall, [[maybe_unused]] test_component& comp)
 {
     syscall.remove_component<test_component>(id);
 }
 
 TEST(ECS, system_defer_syscall_test)
 {
-    using TestResource = System::ResourceManager<1000, test_component>;
-    using TestSyscall = System::Syscall<1000, test_component>;
+    using TestResource = ResourceManager<1000, test_component>;
+    using TestSyscall = Syscall<1000, test_component>;
     TestResource resource;
     TestSyscall syscall{resource};
-    // remove component goes first, but is deferred until exec() triggers
-    System::TaskManager<TestResource, TestSyscall, test_system_3, test_system_2, test_system> task_manager(resource, syscall);
-    const System::pid id_1 = resource.add_process();
+    // the remove component goes first but is deferred until exec() triggers
+    TaskManager<TestResource, TestSyscall, test_system_3, test_system_2, test_system> task_manager(resource, syscall);
+    const pid id_1 = resource.add_process();
     resource.add_resource<test_component>(id_1, test_component{1});
     EXPECT_TRUE(resource.query<test_component>().has(id_1));
     task_manager.run_all();
